+++ src/main/kotlin/internals/instructions/arithmetic/mod.kt
@@ -15,8 +15,8 @@ import returnRegisters
  * @throws GeneralArithmeticException If an arithmetic error occurs during the modulo operation.
  */
 fun Arithmetic.mod(registerA: SuperRegisterType, registerB: SuperRegisterType): Unit = try {
-    val A: Long = registerRead(register = registerA)
-    val B: Long = registerRead(register = registerB)
+    val A: Long = registerRead(register = registerA) as Long
+    val B: Long = registerRead(register = registerB) as Long
     returnRegisters.run {
         write(
             registers = R4,
